mediahal: CF1 Compile the 64 bit library.  [1/1]
PD#SWPL-70993

Problem:
Compile the 64 bit library.

Solution:
Compile the 64 bit library.

Verify:
SC2

SourceCodec:
V3.2.190-gdb7cc9c
db7cc9c mediahal: CF1 Compile the 64 bit library.  [1/1]
6ca0aba MediaSync: CB2 AV sometimes stuck when trickplay change to normal play.
66545f7 tsplayer: CF1 fix video block when surface visibleRegion is 0 [3/3]
e7e9fe2 TsPlayer: CF1 Frame info not right after switch by fcc. [1/1]
7a08efd tsplayer: CB2 send dataloss not need wait avsync. [1/1]
a459110 videodec: CB1 fix the hevc/avc stream head error. [1/1]
8cd8bc4 videodec: CF1 add avs2 support [3/4]
1617808 MediaSync: CF1 AV always stuck. [1/2]
3a47352 Media_sync: CB1 Add startthershold when audio reset refclock. [1/1]
f4c79ae videodec: CB1 fix seed to end error [1/1]

+++ example/AmTsPlayerExample/AmTsPlayerExample.cpp
@@ -14,7 +14,7 @@
 #include <sys/time.h>
 #include <cstdlib>
 #include <time.h>
-#include <sys/types.h>
+//#include <sys/types.h>
 #include <sys/stat.h>
 #include <fcntl.h>
 #include <thread>
@@ -454,13 +454,13 @@ int main(int argc, char **argv)
        file.seekg(0, file.end);
        fsize = file.tellg();
        if (fsize <= 0) {
-           printf("file %s size %lld return\n", inputTsName.c_str(), fsize);
+           printf("file %s size %lld return\n", inputTsName.c_str(),(long long)fsize);
            return 0;
        }
        file.seekg(0, file.beg);
     }
     printf("file name = %s, is_open %d, size %lld, tsType %d\n",
-                inputTsName.c_str(), file.is_open(), fsize, tsType);
+                inputTsName.c_str(), file.is_open(),(long long) fsize, tsType);
 
     //am_tsplayer_handle session;
     am_tsplayer_init_params parm = {tsType, drmmode, 0, 0};
